chg: rename loop_duration to dt, remove the creation of a surface in a phase as it is then blitted on the screen. Move the responsability to have the make_surface and the get_surface methods into childs.
